CHERI: Separate ptr type and index sizes, relate layout to address space
This commit modifies DataLayout in order to allow specification of
multiple different address spaces for different purposes. Each address
space can specify a different pointer layout. This pointer layout can
also consist of a different pointer type size (IE how much space the
pointer itself takes up in memory) and index size (IE the bits available
for addressing).

To obtain this information from the datalayout then we must
differentiate whether we want the type size or the index size. We should
also pass in the address space in which the pointer will exist.