CHERI: Add transmute_size to layout, use for conversion checks
This patch adds a transmute_size to a type layout, which means we can
then use this to determine if the bits of one type can be represented
with the bits of another - IE for converting pointers to/from usize.

Along with changing SizeSkeleton to use this metric for determining if
things are of 'equal' size, we also modify the code emitted for
transmuting a usize to a pointer, such that we are explicitly modifying
the address of a null pointer. Which results in a non-dereferencable
pointer if we are using CHERI.