CHERI: Add range to AllocRange
Sometimes an AllocRange holds a single pointer and in this case if we
use its size as a means of determining the range of addresses it can
represent, we get the wrong answer for CHERI capabilities. So we add a
range field to AllocRange as an equivalent for the size of the value
that can be represented rather than the size of the type itself.