CHERI: Add range to ScalarInt
This change adds a range field to ScalarInt. This is really a hack since
we should only ever need a range field when dealing with pointers, but
at certain points we seem to need to force a value to be an integer,
like when reading the discriminant tag value from a niche representation
when the niche type is a pointer. Thus we need to preserve the pointer
range to perform a valid conversion between it and the integer type we
chose to represent it (IE usize). The reason this is a hack is because
the conversion which takes place is not simply a case of discarding the
upper representation bits, so we actually end up with the wrong value in
the integer destination.