CHERI: Separate total type size versus value size for AllocRange, Primitive, Scalar
Sometimes an AllocRange holds a single pointer and in this case if we
use its size as a means of determining the range of addresses it can
represent, we get the wrong answer for CHERI capabilities. So we split
this into total_size and val_size, where total_size is the size of the
entire range, and val_size is the size of the value that can be
represented within that range - IE for a single pointer, this will be
different to the total size.

We also need to do the same for the Primitive and Scalar types so that
these can report what their range of values is based on their underlying
types rather than just reporting the total type size.