CHERI: Use 'size of value' to tell if alloc range is initialized
This is another workaround for an issue that occurred when the
interpreter needed to query if a scalar capability was fully initialized
or not. Because the reported size of the defined bits was only half of
the total 'allocated range' size for the capability, it was reported
that the scalar was not initialized. This patch ensures we check the
size of the representable value in the allocated range against the
length of bits we know are initialized. This is still a hack because the
bits representing the value in a CHERI capability are not necessarily a
set of contiguous bits.